Feat: Link each button with compatible CameraFrameView
