+++ podong4cuts/podong4cuts/src/Team1_Map/View/MapViewGroup/DetailView.swift
@@ -10,6 +10,7 @@ import SwiftUI
 import CoreLocation
 
 struct DetailView: View {
+    @Environment(\.dismiss) var dismiss
     
     //property
     @ObservedObject var VM: PodongViewModel
@@ -17,7 +18,8 @@ struct DetailView: View {
     
     var selectedNumber: Int = 0
    
-    @State private var showDefaultCameraFrameView = false
+    @Binding var showDefaultCameraFrameView: Bool
+    @Binding var cameraFrameNumber: Int
     
     var body: some View {
         VStack{
@@ -200,13 +202,18 @@ struct DetailView: View {
                                 }
                             }
                             .onTapGesture {
-                                
+                                cameraFrameNumber = selectedNumber
+                                dismiss()
+                                showDefaultCameraFrameView = true
                             }
                         }
                 }//】 Button
                 .padding(.bottom, 5)
                 .vBottom()
             }//】 ZStack
+            NavigationLink("",isActive: $showDefaultCameraFrameView) {
+                DefaultCameraFrameView(selected: selectedNumber)
+            }
         }//: VStack
     }//: Body
     
@@ -238,28 +245,10 @@ struct DetailView: View {
     }
 }
 
-struct DetailView_Previews: PreviewProvider {
-    static var previews: some View {
-        DetailView(VM: PodongViewModel(), selectedNumber: 0)
+//struct DetailView_Previews: PreviewProvider {
+//    static var previews: some View {
+//        DetailView(VM: PodongViewModel(), selectedNumber: 0, showDefaultCameraFrameView: 3)
+//        DetailView(VM: PodongViewModel(), showDefaultCameraFrameView: <#T##Binding<Bool>#>, cameraFrameNumber: <#T##Binding<Int>#>)
            
-    }
-}
-
-
-struct DefaultCameraFrameView: View {
-    let selected: Int
-    @ViewBuilder var body: some View {
-        switch (selected) {
-        case 0:
-            SpaceWalkSpaceCameraFrameView()
-        case 1:
-            TomatiloMustacheCameraFrameView()
-        case 2:
-            YeongilgyoKrabFilterView()
-        case 3:
-            WhaleTailCameraFrameView()
-        default:
-            WhaleTailCameraFrameView()
-        }
-    }
-}
+//    }
+//}
