+++ podong4cuts/podong4cuts/src/Team2_AR/View/CameraFrameViews/CameraFrameView.swift
@@ -133,7 +133,7 @@ struct CameraFrameView: View {
                 }
                 
                 NavigationLink("", isActive: $showPhotoLibrary) {
-                    PhotoLibraryView(arViewContainer: arViewContainer)
+                    PhotoLibraryView()
                 }
             }
         }
